refactor: Refactor PostTopBox LeftArea Component
